+++ com.unity.render-pipelines.high-definition/Runtime/PostProcessing/PostProcessSystem.cs
@@ -19,7 +19,7 @@ namespace UnityEngine.Rendering.HighDefinition
             NeighborhoodBlending = 2
         }
 
-        const GraphicsFormat k_ColorFormat         = GraphicsFormat.B10G11R11_UFloatPack32;
+        GraphicsFormat m_ColorFormat               = GraphicsFormat.B10G11R11_UFloatPack32;
         const GraphicsFormat k_CoCFormat           = GraphicsFormat.R16_SFloat;
         const GraphicsFormat k_ExposureFormat      = GraphicsFormat.R32G32_SFloat;
 
@@ -108,7 +108,12 @@ namespace UnityEngine.Rendering.HighDefinition
         RTHandle m_TempTexture1024; // RGHalf
         RTHandle m_TempTexture32;   // RGHalf
 
-        readonly bool m_KeepAlpha;
+        // HDRP has the following behavior regarding alpha:
+        // - If post processing is disabled, the alpha channel of the rendering passes (if any) will be passed to the frame buffer by the final pass
+        // - If post processing is enabled, then post processing passes will either copy (exposure, color grading, etc) or process (DoF, TAA, etc) the alpha channel, if one exists.
+        // If the user explicitly requests a color buffer without alpha for post-processing (for performance reasons) but the rendering passes have alpha, then the alpha will be copied.
+        readonly bool m_EnableAlpha;
+        readonly bool m_KeepAlpha; 
         RTHandle m_AlphaTexture; // RHalf
 
         readonly TargetPool m_Pool;
@@ -158,6 +163,18 @@ namespace UnityEngine.Rendering.HighDefinition
             PushUberFeature(UberPostFeatureFlags.ChromaticAberration | UberPostFeatureFlags.LensDistortion);
             PushUberFeature(UberPostFeatureFlags.Vignette | UberPostFeatureFlags.LensDistortion);
             PushUberFeature(UberPostFeatureFlags.ChromaticAberration | UberPostFeatureFlags.Vignette | UberPostFeatureFlags.LensDistortion);
+
+            //Alpha mask variants:
+            {
+                PushUberFeature(UberPostFeatureFlags.EnableAlpha);
+                PushUberFeature(UberPostFeatureFlags.ChromaticAberration | UberPostFeatureFlags.EnableAlpha);
+                PushUberFeature(UberPostFeatureFlags.Vignette | UberPostFeatureFlags.EnableAlpha);
+                PushUberFeature(UberPostFeatureFlags.LensDistortion | UberPostFeatureFlags.EnableAlpha);
+                PushUberFeature(UberPostFeatureFlags.ChromaticAberration | UberPostFeatureFlags.Vignette | UberPostFeatureFlags.EnableAlpha);
+                PushUberFeature(UberPostFeatureFlags.ChromaticAberration | UberPostFeatureFlags.LensDistortion | UberPostFeatureFlags.EnableAlpha);
+                PushUberFeature(UberPostFeatureFlags.Vignette | UberPostFeatureFlags.LensDistortion | UberPostFeatureFlags.EnableAlpha);
+                PushUberFeature(UberPostFeatureFlags.ChromaticAberration | UberPostFeatureFlags.Vignette | UberPostFeatureFlags.LensDistortion | UberPostFeatureFlags.EnableAlpha);
+            }
 
             // Grading specific
             m_HableCurve = new HableCurve();
@@ -213,17 +230,28 @@ namespace UnityEngine.Rendering.HighDefinition
                 32, 32, colorFormat: GraphicsFormat.R16G16_SFloat,
                 enableRandomWrite: true, name: "Average Luminance Temp 32"
             );
-
-            // Used to copy the alpha channel of the color buffer if the format is set to fp16
-            m_KeepAlpha = hdAsset.currentPlatformRenderPipelineSettings.keepAlpha;
-            if (m_KeepAlpha)
+   
+            m_ColorFormat = (GraphicsFormat)hdAsset.currentPlatformRenderPipelineSettings.postProcessSettings.bufferFormat;
+            m_EnableAlpha = false;
+            m_KeepAlpha = false;
+            if(m_PostProcessEnabled)
             {
-                m_AlphaTexture = RTHandles.Alloc(
-                    Vector2.one, slices: TextureXR.slices, dimension: TextureXR.dimension,
-                    colorFormat: GraphicsFormat.R16_SFloat, enableRandomWrite: true, name: "Alpha Channel Copy"
-                );
+                // if both rendering and post-processing support an alpha channel, then post-processing will process (or copy) the alpha
+                m_EnableAlpha = hdAsset.currentPlatformRenderPipelineSettings.supportsAlpha && hdAsset.currentPlatformRenderPipelineSettings.postProcessSettings.supportsAlpha;
+            }
+            if(m_EnableAlpha == false)
+            {
+                // if only rendering has an alpha channel (and not post-processing), then we just copy the alpha to the output (but we don't process it).
+                m_KeepAlpha = hdAsset.currentPlatformRenderPipelineSettings.supportsAlpha;
             }
 
+            if(m_KeepAlpha)
+            {
+                m_AlphaTexture = RTHandles.Alloc(
+                   Vector2.one, slices: TextureXR.slices, dimension: TextureXR.dimension,
+                   colorFormat: GraphicsFormat.R16_SFloat, enableRandomWrite: true, name: "Alpha Channel Copy"
+               );
+            }
             ResetHistory();
         }
 
@@ -363,15 +391,14 @@ namespace UnityEngine.Rendering.HighDefinition
 
             using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.PostProcessing)))
             {
-                // Save the alpha and apply it back into the final pass if working in fp16
-                if (m_KeepAlpha)
+            	// Save the alpha and apply it back into the final pass if rendering in fp16 and post-processing in r11g11b10
+                if(m_KeepAlpha)
                 {
                     using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.AlphaCopy)))
                     {
                         DoCopyAlpha(cmd, camera, colorBuffer);
                     }
                 }
-
                 var source = colorBuffer;
 
                 if (m_PostProcessEnabled)
@@ -406,7 +433,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     {
                         using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.StopNaNs)))
                         {
-                            var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                            var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                             DoStopNaNs(cmd, camera, source, destination);
                             PoolSource(ref source, destination);
                         }
@@ -426,7 +453,7 @@ namespace UnityEngine.Rendering.HighDefinition
                         // near 0
                         if (m_ResetHistory)
                         {
-                            var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                            var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
 
                             var cs = m_Resources.shaders.applyExposureCS;
                             int kernel = cs.FindKernel("KMain");
@@ -458,7 +485,7 @@ namespace UnityEngine.Rendering.HighDefinition
                         {
                             using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.TemporalAntialiasing)))
                             {
-                                var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                                var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                                 DoTemporalAntialiasing(cmd, camera, source, destination, depthBuffer);
                                 PoolSource(ref source, destination);
                             }
@@ -467,7 +494,7 @@ namespace UnityEngine.Rendering.HighDefinition
                         {
                             using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.SMAA)))
                             {
-                                var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                                var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                                 DoSMAA(cmd, camera, source, destination, depthBuffer);
                                 PoolSource(ref source, destination);
                             }
@@ -489,7 +516,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     {
                         using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.DepthOfField)))
                         {
-                            var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                            var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                             DoDepthOfField(cmd, camera, source, destination, taaEnabled);
                             PoolSource(ref source, destination);
                         }
@@ -501,7 +528,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     {
                         using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.MotionBlur)))
                         {
-                            var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                            var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                             DoMotionBlur(cmd, camera, source, destination);
                             PoolSource(ref source, destination);
                         }
@@ -515,7 +542,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     {
                         using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.PaniniProjection)))
                         {
-                            var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                            var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                             DoPaniniProjection(cmd, camera, source, destination);
                             PoolSource(ref source, destination);
                         }
@@ -559,7 +586,7 @@ namespace UnityEngine.Rendering.HighDefinition
                         DoVignette(cmd, cs, kernel, featureFlags);
 
                         // Run
-                        var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                        var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
 
                         bool outputColorLog = m_HDInstance.m_CurrentDebugDisplaySettings.data.fullScreenDebugMode == FullScreenDebugMode.ColorLog;
                         cmd.SetComputeVectorParam(cs, "_DebugFlags", new Vector4(outputColorLog ? 1 : 0, 0, 0, 0));
@@ -591,7 +618,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 {
                     using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.FXAA)))
                     {
-                        var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                        var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                         DoFXAA(cmd, camera, source, destination);
                         PoolSource(ref source, destination);
                     }
@@ -603,7 +630,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 {
                     using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.ContrastAdaptiveSharpen)))
                     {
-                        var destination = m_Pool.Get(Vector2.one , k_ColorFormat);
+                        var destination = m_Pool.Get(Vector2.one , m_ColorFormat);
                         
                         var cs = m_Resources.shaders.contrastAdaptiveSharpenCS;
                         int kInit = cs.FindKernel("KInitialize");
@@ -672,6 +699,11 @@ namespace UnityEngine.Rendering.HighDefinition
 
             if (m_LensDistortion.IsActive() && !isSceneView && m_LensDistortionFS)
                 flags |= UberPostFeatureFlags.LensDistortion;
+
+            if (m_EnableAlpha)
+            {
+                flags |= UberPostFeatureFlags.EnableAlpha;
+            }
 
             return flags;
         }
@@ -888,6 +920,10 @@ namespace UnityEngine.Rendering.HighDefinition
         {
             GrabTemporalAntialiasingHistoryTextures(camera, out var prevHistory, out var nextHistory);
 
+            if (m_EnableAlpha)
+            {
+                m_TemporalAAMaterial.EnableKeyword("ENABLE_ALPHA");
+            }
 
             if (m_ResetHistory)
             {
@@ -913,13 +949,13 @@ namespace UnityEngine.Rendering.HighDefinition
             cmd.ClearRandomWriteTargets();
         }
 
-        static void GrabTemporalAntialiasingHistoryTextures(HDCamera camera, out RTHandle previous, out RTHandle next)
+        void GrabTemporalAntialiasingHistoryTextures(HDCamera camera, out RTHandle previous, out RTHandle next)
         {
             RTHandle Allocator(string id, int frameIndex, RTHandleSystem rtHandleSystem)
             {
                 return rtHandleSystem.Alloc(
                     Vector2.one, TextureXR.slices, DepthBits.None, dimension: TextureXR.dimension,
-                    filterMode: FilterMode.Bilinear, colorFormat: k_ColorFormat,
+                    filterMode: FilterMode.Bilinear, colorFormat: m_ColorFormat,
                     enableRandomWrite: true, useDynamicScale: true, name: "TAA History"
                 );
             }
@@ -932,7 +968,6 @@ namespace UnityEngine.Rendering.HighDefinition
         #endregion
 
         #region Depth Of Field
-
         //
         // Reference used:
         //   "A Lens and Aperture Camera Model for Synthetic Image Generation" [Potmesil81]
@@ -993,6 +1028,10 @@ namespace UnityEngine.Rendering.HighDefinition
             float resolutionScale = (camera.actualHeight / 1080f) * (scale * 2f);
             int farSamples = Mathf.CeilToInt(m_DepthOfField.farSampleCount * resolutionScale);
             int nearSamples = Mathf.CeilToInt(m_DepthOfField.nearSampleCount * resolutionScale);
+            // We want at least 3 samples for both far and near
+            farSamples = Mathf.Max(3, farSamples);
+            nearSamples = Mathf.Max(3, nearSamples);
+
             float farMaxBlur = m_DepthOfField.farMaxBlur * resolutionScale;
             float nearMaxBlur = m_DepthOfField.nearMaxBlur * resolutionScale;
 
@@ -1012,8 +1051,8 @@ namespace UnityEngine.Rendering.HighDefinition
 
             if (nearLayerActive)
             {
-                pingNearRGB = m_Pool.Get(screenScale, k_ColorFormat);
-                pongNearRGB = m_Pool.Get(screenScale, k_ColorFormat);
+                pingNearRGB = m_Pool.Get(screenScale, m_ColorFormat);
+                pongNearRGB = m_Pool.Get(screenScale, m_ColorFormat);
                 nearCoC = m_Pool.Get(screenScale, k_CoCFormat);
                 nearAlpha = m_Pool.Get(screenScale, k_CoCFormat);
                 dilatedNearCoC = m_Pool.Get(screenScale, k_CoCFormat);
@@ -1021,8 +1060,8 @@ namespace UnityEngine.Rendering.HighDefinition
 
             if (farLayerActive)
             {
-                pingFarRGB = m_Pool.Get(screenScale, k_ColorFormat, true);
-                pongFarRGB = m_Pool.Get(screenScale, k_ColorFormat);
+                pingFarRGB = m_Pool.Get(screenScale, m_ColorFormat, true);
+                pongFarRGB = m_Pool.Get(screenScale, m_ColorFormat);
                 farCoC = m_Pool.Get(screenScale, k_CoCFormat, true);
             }
 
@@ -1078,7 +1117,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     float F = camera.camera.focalLength / 1000f;
                     float A = camera.camera.focalLength / m_PhysicalCamera.aperture;
                     float P = m_DepthOfField.focusDistance.value;
-                    float maxCoC = (A * F) / (P - F);
+                    float maxCoC = (A * F) / Mathf.Max((P - F), 1e-6f);
 
                     kernel = cs.FindKernel("KMainPhysical");
                     cmd.SetComputeVectorParam(cs, HDShaderIDs._Params, new Vector4(P, maxCoC, 0f, 0f));
@@ -1088,7 +1127,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     float nearEnd = m_DepthOfField.nearFocusEnd.value;
                     float nearStart = Mathf.Min(m_DepthOfField.nearFocusStart.value, nearEnd - 1e-5f);
                     float farStart = Mathf.Max(m_DepthOfField.farFocusStart.value, nearEnd);
-                    float farEnd = Mathf.Max(m_DepthOfField.farFocusEnd.value, farStart);
+                    float farEnd = Mathf.Max(m_DepthOfField.farFocusEnd.value, farStart + 1e-5f);
 
                     kernel = cs.FindKernel("KMainManual");
                     cmd.SetComputeVectorParam(cs, HDShaderIDs._Params, new Vector4(nearStart, nearEnd, farStart, farEnd));
@@ -1132,9 +1171,20 @@ namespace UnityEngine.Rendering.HighDefinition
 
                 cs = m_Resources.shaders.depthOfFieldPrefilterCS;
 
-                kernel = m_DepthOfField.resolution == DepthOfFieldResolution.Full
-                    ? cs.FindKernel(bothLayersActive ? "KMainNearFarFullRes" : nearLayerActive ? "KMainNearFullRes" : "KMainFarFullRes")
-                    : cs.FindKernel(bothLayersActive ? "KMainNearFar" : nearLayerActive ? "KMainNear" : "KMainFar");
+                if(m_EnableAlpha)
+                {
+                    // kernels with alpha channel:
+                    kernel = cs.FindKernel(m_DepthOfField.resolution == DepthOfFieldResolution.Full ?
+                    (bothLayersActive ? "KMainNearFarFullResAlpha" : nearLayerActive ? "KMainNearFullResAlpha" : "KMainFarFullResAlpha") :
+                    (bothLayersActive ? "KMainNearFarAlpha" : nearLayerActive ? "KMainNearAlpha" : "KMainFarAlpha"));
+                }
+                else
+                {
+                    // kernels without alpha channel:
+                    kernel = cs.FindKernel(m_DepthOfField.resolution == DepthOfFieldResolution.Full ?
+                    (bothLayersActive ? "KMainNearFarFullRes" : nearLayerActive ? "KMainNearFullRes" : "KMainFarFullRes") :
+                    (bothLayersActive ? "KMainNearFar" : nearLayerActive ? "KMainNear" : "KMainFar"));
+                }
 
                 cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputTexture, source);
                 cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputCoCTexture, fullresCoC);
@@ -1171,14 +1221,14 @@ namespace UnityEngine.Rendering.HighDefinition
                     {
                         // The other compute fails hard on Intel because of texture format issues
                         cs = m_Resources.shaders.depthOfFieldMipSafeCS;
-                        kernel = cs.FindKernel("KMain");
+                        kernel = cs.FindKernel(m_EnableAlpha? "KMainAlpha" : "KMain");
                         var mipScale = scale;
 
                         for (int i = 0; i < 4; i++)
                         {
                             mipScale *= 0.5f;
                             var size = new Vector2Int(Mathf.RoundToInt(camera.actualWidth * mipScale), Mathf.RoundToInt(camera.actualHeight * mipScale));
-                            var mip = m_Pool.Get(new Vector2(mipScale, mipScale), k_ColorFormat);
+                            var mip = m_Pool.Get(new Vector2(mipScale, mipScale), m_ColorFormat);
 
                             cmd.SetComputeVectorParam(cs, HDShaderIDs._TexelSize, new Vector4(size.x, size.y, 1f / size.x, 1f / size.y));
                             int gx = (size.x + 7) / 8;
@@ -1200,7 +1250,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     else
                     {
                         cs = m_Resources.shaders.depthOfFieldMipCS;
-                        kernel = cs.FindKernel("KMainColor");
+                        kernel = cs.FindKernel(m_EnableAlpha ? "KMainColorAlpha": "KMainColor");
                         cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputTexture, pingFarRGB, 0);
                         cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._OutputMip1, pingFarRGB, 1);
                         cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._OutputMip2, pingFarRGB, 2);
@@ -1314,7 +1364,10 @@ namespace UnityEngine.Rendering.HighDefinition
                     }
 
                     cs = m_Resources.shaders.depthOfFieldGatherCS;
-                    kernel = cs.FindKernel(useTiles ? "KMainFarTiles" : "KMainFar");
+                    kernel = m_EnableAlpha ?
+                        cs.FindKernel(useTiles ? "KMainFarTilesAlpha" : "KMainFarAlpha") :
+                        cs.FindKernel(useTiles ? "KMainFarTiles" : "KMainFar");
+
                     cmd.SetComputeVectorParam(cs, HDShaderIDs._Params, new Vector4(farSamples, farSamples * farSamples, barrelClipping, farMaxBlur));
                     cmd.SetComputeVectorParam(cs, HDShaderIDs._TexelSize, new Vector4(targetWidth, targetHeight, 1f / targetWidth, 1f / targetHeight));
                     cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputTexture, pingFarRGB);
@@ -1345,7 +1398,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     if (farLayerActive)
                     {
                         cs = m_Resources.shaders.depthOfFieldCombineCS;
-                        kernel = cs.FindKernel("KMainPreCombineFar");
+                        kernel = cs.FindKernel(m_EnableAlpha ? "KMainPreCombineFarAlpha" : "KMainPreCombineFar");
                         cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputTexture, pingNearRGB);
                         cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputFarTexture, pongFarRGB);
                         cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputCoCTexture, farCoC);
@@ -1375,7 +1428,9 @@ namespace UnityEngine.Rendering.HighDefinition
                     }
 
                     cs = m_Resources.shaders.depthOfFieldGatherCS;
-                    kernel = cs.FindKernel(useTiles ? "KMainNearTiles" : "KMainNear");
+                    kernel = m_EnableAlpha ?
+                        cs.FindKernel(useTiles ? "KMainNearTilesAlpha" : "KMainNearAlpha") :
+                        cs.FindKernel(useTiles ? "KMainNearTiles" : "KMainNear");
                     cmd.SetComputeVectorParam(cs, HDShaderIDs._Params, new Vector4(nearSamples, nearSamples * nearSamples, barrelClipping, nearMaxBlur));
                     cmd.SetComputeVectorParam(cs, HDShaderIDs._TexelSize, new Vector4(targetWidth, targetHeight, 1f / targetWidth, 1f / targetHeight));
                     cmd.SetComputeTextureParam(cs, kernel, HDShaderIDs._InputTexture, pingNearRGB);
@@ -1404,12 +1459,24 @@ namespace UnityEngine.Rendering.HighDefinition
 
                 cs = m_Resources.shaders.depthOfFieldCombineCS;
 
-                if (m_DepthOfField.resolution == DepthOfFieldResolution.Full)
-                    kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarFullRes" : nearLayerActive ? "KMainNearFullRes" : "KMainFarFullRes");
-                else if (hqFiltering)
-                    kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarHighQ" : nearLayerActive ? "KMainNearHighQ" : "KMainFarHighQ");
+                if(m_EnableAlpha)
+                {
+                    if (m_DepthOfField.resolution == DepthOfFieldResolution.Full)
+                        kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarFullResAlpha" : nearLayerActive ? "KMainNearFullResAlpha" : "KMainFarFullResAlpha");
+                    else if (hqFiltering)
+                        kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarHighQAlpha" : nearLayerActive ? "KMainNearHighQAlpha" : "KMainFarHighQAlpha");
+                    else
+                        kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarLowQAlpha" : nearLayerActive ? "KMainNearLowQAlpha" : "KMainFarLowQAlpha");
+                }
                 else
-                    kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarLowQ" : nearLayerActive ? "KMainNearLowQ" : "KMainFarLowQ");
+                {
+                    if (m_DepthOfField.resolution == DepthOfFieldResolution.Full)
+                        kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarFullRes" : nearLayerActive ? "KMainNearFullRes" : "KMainFarFullRes");
+                    else if (hqFiltering)
+                        kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarHighQ" : nearLayerActive ? "KMainNearHighQ" : "KMainFarHighQ");
+                    else
+                        kernel = cs.FindKernel(bothLayersActive ? "KMainNearFarLowQ" : nearLayerActive ? "KMainNearLowQ" : "KMainFarLowQ");
+                }
 
                 if (nearLayerActive)
                 {
@@ -1799,8 +1866,8 @@ namespace UnityEngine.Rendering.HighDefinition
                 var pixelSize = new Vector2Int(pw, ph);
 
                 mipSizes[i] = pixelSize;
-                m_BloomMipsDown[i] = m_Pool.Get(scale, k_ColorFormat);
-                m_BloomMipsUp[i] = m_Pool.Get(scale, k_ColorFormat);
+                m_BloomMipsDown[i] = m_Pool.Get(scale, m_ColorFormat);
+                m_BloomMipsUp[i] = m_Pool.Get(scale, m_ColorFormat);
             }
 
             // All the computes for this effect use the same group size so let's use a local
@@ -2418,6 +2485,11 @@ namespace UnityEngine.Rendering.HighDefinition
                 m_FinalPassMaterial.SetFloat(HDShaderIDs._KeepAlpha, 0.0f);
             }
 
+            if (m_EnableAlpha)
+            {
+                m_FinalPassMaterial.EnableKeyword("ENABLE_ALPHA");
+            }
+
             m_FinalPassMaterial.SetVector(HDShaderIDs._UVTransform,
                 flipY
                 ? new Vector4(1.0f, -1.0f, 0.0f, 1.0f)
@@ -2495,7 +2567,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 {
                     if (camera.camera.cameraType != CameraType.SceneView || customPP.visibleInSceneView)
                     {
-                        var destination = m_Pool.Get(Vector2.one, k_ColorFormat);
+                        var destination = m_Pool.Get(Vector2.one, m_ColorFormat);
                         CoreUtils.SetRenderTarget(cmd, destination);
                         {
                             cmd.BeginSample(customPP.name);